Fixes :   shows movie title nice.   cleans code
