+++ app/src/main/java/com/example/tmdb_movies/ui/screens/HomeScreen.kt
@@ -114,97 +114,6 @@ fun HomeScreen(
 
 
 @Composable
-fun MovieCategoryRow(
-    title: String,
-    movies: List<Movie>,
-    detailViewModel: DetailViewModel,
-    cardClicked: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Column(modifier = modifier.padding(start = 8.dp)) {
-        Text(text = title, fontSize = 22.sp, modifier = Modifier.padding(start = 10.dp))
-        LazyRow(
-            modifier = Modifier
-                .padding(start = 10.dp, end = 10.dp)
-                .height(250.dp)
-        ) {
-            items(items = movies, key = { movie -> movie.id }) { movie ->
-                MovieCard(
-                    movie = movie,
-                    detailViewModel = detailViewModel,
-                    cardClicked = cardClicked,
-                    modifier = Modifier
-                        .padding(4.dp)
-                        .fillMaxHeight()
-                        .width(170.dp)
-                )
-            }
-        }
-    }
-}
-
-
-@Composable
-fun LoadingScreen(modifier: Modifier = Modifier) {
-    Image(
-        modifier = modifier.size(200.dp),
-        painter = painterResource(R.drawable.loading_img),
-        contentDescription = "Loading"
-    )
-}
-
-@Composable
-fun ErrorScreen(retryAction: () -> Unit, modifier: Modifier = Modifier) {
-    Column(
-        modifier = modifier,
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Image(
-            painter = painterResource(id = R.drawable.ic_connection_error), contentDescription = ""
-        )
-        Text(text = "Failed to load", modifier = Modifier.padding(16.dp))
-        Button(onClick = retryAction) {
-            Text("retry")
-        }
-    }
-}
-
-@Composable
-fun MovieCard(
-    movie: Movie,
-    detailViewModel: DetailViewModel,
-    cardClicked: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    var imagePainter: Painter? by remember { mutableStateOf(null) }
-
-    Card(
-        modifier = modifier,
-        shape = MaterialTheme.shapes.medium,
-        elevation = CardDefaults.cardElevation(defaultElevation = 8.dp),
-        onClick = {
-            detailViewModel.setCardDetail(movie, imagePainter)
-            cardClicked()
-        },
-    ) {
-        AsyncImage(
-            model = movie.fullPosterUrl,
-            error = painterResource(R.drawable.ic_broken_image),
-            placeholder = painterResource(R.drawable.loading_img),
-            contentDescription = null,
-            contentScale = ContentScale.Crop,
-            modifier = Modifier.fillMaxSize(),
-            onSuccess = { painter ->
-                imagePainter = painter.painter
-            }
-        )
-        Text(text = movie.title)
-    }
-}
-
-
-@Composable
 private fun BottomNavigationBar(
     currentTab: AppType,
     onTabPressed: (AppType) -> Unit,
